requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
boto3==1.34.131
botocore==1.34.131
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
jmespath==1.0.1
multidict==6.0.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
schedule==1.2.1
six==1.16.0
urllib3==2.2.2
yarl==1.9.4
